# internal/models/other.go
package models

type ErrorResponse struct {
	Error     string `json:"error,omitempty"`
	Code      string `json:"code,omitempty"`
	RequestID string `json:"request_id,omitempty"`
}

type MessageResponse struct {
	Message string `json:"message,omitempty"`
}